+++ src/diagnostics/common.ts
@@ -19,6 +19,7 @@ export enum DiagnosticCode {
     DuplicatedDefinition,
     UndefinedExpression,
     UndefinedStaticMessage,
+    UndefinedMessage,
     UndefinedSymbol,
     UndefinedTask,
     UndefinedAttribute,
@@ -54,7 +55,7 @@ export const Errors = {
     invalidStaticMessageDefinition: (range: Range, id: number, name: string) =>
         makeDiagnostic(range, DiagnosticCode.UndefinedStaticMessage, '\'' + name + '\' is not a valid alias for message ' + id + '.', DiagnosticSeverity.Error),
     undefinedMessage: (range: Range, message: string) =>
-        makeDiagnostic(range, DiagnosticCode.UndefinedExpression, 'Reference to undefined message: ' + message + '.', DiagnosticSeverity.Error),
+        makeDiagnostic(range, DiagnosticCode.UndefinedMessage, 'Reference to undefined message: ' + message + '.', DiagnosticSeverity.Error),
     undefinedSymbol: (range: Range, symbol: string) =>
         makeDiagnostic(range, DiagnosticCode.UndefinedSymbol, 'Reference to undefined symbol: ' + symbol + '.', DiagnosticSeverity.Error),
     undefinedTask: (range: Range, symbol: string) =>